// condoplanner-app/src/apiClient/models/RegisterRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RegisterRequest
 */
export interface RegisterRequest {
    /**
     * 
     * @type {string}
     * @memberof RegisterRequest
     */
    nome?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RegisterRequest
     */
    email?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RegisterRequest
     */
    telefone?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RegisterRequest
     */
    senha?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RegisterRequest
     */
    cpf?: string | null;
}

/**
 * Check if a given object implements the RegisterRequest interface.
 */
export function instanceOfRegisterRequest(value: object): value is RegisterRequest {
    return true;
}

export function RegisterRequestFromJSON(json: any): RegisterRequest {
    return RegisterRequestFromJSONTyped(json, false);
}

export function RegisterRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): RegisterRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'nome': json['nome'] == null ? undefined : json['nome'],
        'email': json['email'] == null ? undefined : json['email'],
        'telefone': json['telefone'] == null ? undefined : json['telefone'],
        'senha': json['senha'] == null ? undefined : json['senha'],
        'cpf': json['cpf'] == null ? undefined : json['cpf'],
    };
}

export function RegisterRequestToJSON(json: any): RegisterRequest {
    return RegisterRequestToJSONTyped(json, false);
}

export function RegisterRequestToJSONTyped(value?: RegisterRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'nome': value['nome'],
        'email': value['email'],
        'telefone': value['telefone'],
        'senha': value['senha'],
        'cpf': value['cpf'],
    };
}